Pgbench Serialization errors and set default transaction isolation level
Now session is not disconnected because of serialization errors. If there was
such error during script execution this "transaction" is marked appropriately in
logs. Number of such errors is printed in progress, in aggregation logs and in
the end with other results (all and for each script). You can set the default
transaction isolation level by the appropriate benchmarking option (-I).