Pgbench Deadlock errors
Now session is not disconnected because of deadlock errors. If there was such
error during script execution this "transaction" is marked appropriately in logs
(if there was serialization error too, both failures are marked). Number of
deadlock errors is printed in progress, in aggregation logs and in the end with
other results (all and for each script).