Pgbench Retry the failed transactions (with tests)
Now the failed transactions can be retried again and again until they end
sucessfully or their number of attempts reaches maximum. You can set the maximum
number of attempts by the appropriate benchmarking option
(--max-attempts-number). Its default value is 1. Statistics of attempts is
printed in progress, in transaction / aggregation logs and in the end with other
results (all and for each script). Also it is printed per commands which start
transactions if you use the appropriate benchmarking option
(--report-per-command, -r).