Pgbench Fix report of failures and retries, rename option for tries
Now the reports include the total number of failures without retrying and the
statistics of retries: the total number of the retried transactions, the total
number of retries and the number of retries after each type of failure
(serialization or deadlock).  Usually the retries and/or failures are printed if
they are not equal to zeros.  In transaction/aggregation logs the failures are
always printed and the retries are printed if max_tries is greater than 1. It is
done for the general format of the log during the execution of the program.

The benchmarking option --max-attempts-number is now called --max-tries.