Pgbench Fix processing of scripts whose last transaction is not completed
There's no error if the last transaction in the script is not completed. But the
transactions started in the previous scripts and/or not ending in the current
script, are not rolled back and retried after the failure. Such script try is
reported as failed because it contains a failure that was not rolled back and
retried.